# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Motion Index Fiber API
  description: |
    High-performance legal document processing API for California public defenders.
    
    Built with Go and Fiber v2, featuring DigitalOcean cloud integration, AI-powered document classification, and comprehensive search capabilities.
    
    ## Features
    - Document upload and processing (PDF, DOCX, TXT, RTF)
    - AI-powered document classification using GPT-4
    - Full-text search with legal metadata filtering
    - DigitalOcean Spaces storage with CDN
    - JWT authentication via Supabase
    - Real-time document processing pipeline
    
    ## Authentication
    All protected endpoints require a valid JWT token in the `Authorization` header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    See the [Authentication Guide](./authentication.md) for detailed information.
    
    ## Error Handling
    The API uses structured error responses with consistent format:
    ```json
    {
      "success": false,
      "error": {
        "code": "ERROR_CODE",
        "message": "Human-readable error message",
        "details": {
          "additional": "context-specific information"
        }
      }
    }
    ```
    
    See the [Error Codes Reference](./error-codes.md) for all possible error codes.
  version: 1.0.0
  contact:
    name: Motion Index Support
    email: support@motion-index.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
...
paths:
...
  /api/v1/search:
    post:
      tags:
        - Search
      summary: Search documents
      description: |
        Search through the legal document database with support for:
        - Full-text search across document content
        - Metadata filtering (court, judge, case type, date ranges)
        - Sorting and pagination
        - Aggregations for faceted search
        - Highlighting of search terms
      operationId: searchDocuments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              simple_search:
                summary: Simple text search
                value:
                  query: "motion to dismiss"
                  limit: 20
                  offset: 0
              filtered_search:
                summary: Search with filters
                value:
                  query: "habeas corpus"
                  filters:
                    court: ["Superior Court of California"]
                    document_type: ["Motion", "Brief"]
                    date_range:
                      start: "2023-01-01"
                      end: "2023-12-31"
                  sort:
                    field: "filing_date"
                    order: "desc"
                  limit: 10
                  offset: 0
                  include_aggregations: true
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          description: Invalid search request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Search service error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/documents/{document_id}:
    get:
      tags:
        - Documents
      summary: Get document by ID
      description: Retrieve document metadata and content information by ID
      operationId: getDocument
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
          description: Unique document identifier
          example: "doc_123456"
      responses:
        '200':
          description: Document information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Documents
      summary: Delete document
      description: Delete a document from storage and search index (requires authentication)
      operationId: deleteDocument
      security:
        - BearerAuth: []
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
          description: Unique document identifier
          example: "doc_123456"
      responses:
        '200':
          description: Document deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/document-stats:
    get:
      tags:
        - Statistics
      summary: Get document statistics
      description: Get overall statistics about the document collection
      operationId: getDocumentStats
      responses:
        '200':
          description: Document statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentStatsResponse'
              example:
                success: true
                data:
                  total_documents: 15432
                  total_size_bytes: 8934567234
                  document_types:
                    Motion: 5678
                    Brief: 3456
                    Order: 2345
                    Judgment: 1234
                    Other: 2719
                  courts:
                    "Superior Court of California": 12345
                    "Court of Appeal": 2345
                    "Supreme Court": 742
                  date_range:
                    earliest: "2020-01-01"
                    latest: "2024-01-15"
                  last_updated: "2024-01-15T10:30:00Z"
                message: "Statistics retrieved successfully"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/metadata-fields:
    get:
      tags:
        - Metadata
      summary: Get available metadata fields
      description: Get list of available metadata fields for filtering and search
      operationId: getMetadataFields
      responses:
        '200':
          description: Available metadata fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MetadataFieldsResponse'
              example:
                success: true
                data:
                  fields:
                    court:
                      type: "string"
                      searchable: true
                      filterable: true
                      values:
                        - "Superior Court of California"
                        - "Court of Appeal"
                        - "Supreme Court of California"
                    document_type:
                      type: "string"
                      searchable: true
                      filterable: true
                      values:
                        - "Motion"
                        - "Brief"
                        - "Order"
                        - "Judgment"
                    filing_date:
                      type: "date"
                      searchable: false
                      filterable: true
                      range:
                        min: "2020-01-01"
                        max: "2024-01-15"
                    case_number:
                      type: "string"
                      searchable: true
                      filterable: true
                message: "Metadata fields retrieved successfully"
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from Supabase authentication.
        
        Include in requests as: `Authorization: Bearer <your-jwt-token>`
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data (structure varies by endpoint)
        message:
          type: string
          description: Human-readable success message
          example: "Operation completed successfully"
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid input parameters"
            details:
              type: object
              description: Additional error context and details
              additionalProperties: true
          required:
            - code
            - message
      required:
        - success
        - error
...
    SearchRequest:
      type: object
      properties:
        query:
          type: string
          description: Search query text
          example: "motion to dismiss"
        filters:
          type: object
          properties:
            court:
              type: array
              items:
                type: string
              example: ["Superior Court of California"]
            judge:
              type: array
              items:
                type: string
              example: ["Hon. Jane Smith"]
            document_type:
              type: array
              items:
                type: string
              example: ["Motion", "Brief"]
            category:
              type: array
              items:
                type: string
              example: ["Pretrial Motion"]
            case_number:
              type: array
              items:
                type: string
              example: ["CV-2024-001"]
            date_range:
              type: object
              properties:
                start:
                  type: string
                  format: date
                  example: "2023-01-01"
                end:
                  type: string
                  format: date
                  example: "2023-12-31"
        sort:
          type: object
          properties:
            field:
              type: string
              enum: [relevance, filing_date, case_number, document_type]
              default: relevance
            order:
              type: string
              enum: [asc, desc]
              default: desc
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
          description: Number of results to return
        offset:
          type: integer
          minimum: 0
          default: 0
          description: Number of results to skip
        include_aggregations:
          type: boolean
          default: false
          description: Include aggregations for faceted search
        include_highlights:
          type: boolean
          default: true
          description: Include highlighted search terms
    SearchResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total:
                  type: integer
                  description: Total number of matching documents
                  example: 1247
                results:
                  type: array
                  items:
                    $ref: '#/components/schemas/DocumentResult'
                aggregations:
                  type: object
                  description: Search aggregations (if requested)
                  properties:
                    document_types:
                      type: array
                      items:
                        type: object
                        properties:
                          key:
                            type: string
                          doc_count:
                            type: integer
                    courts:
                      type: array
                      items:
                        type: object
                        properties:
                          key:
                            type: string
                          doc_count:
                            type: integer
                query_time_ms:
                  type: integer
                  description: Query execution time in milliseconds
                  example: 23
              required:
                - total
                - results
    DocumentResult:
      type: object
      properties:
        document_id:
          type: string
          example: "doc_123456"
        filename:
          type: string
          example: "motion_to_dismiss.pdf"
        title:
          type: string
          example: "Motion to Dismiss - Lack of Jurisdiction"
        content_preview:
          type: string
          description: First 200 characters of document content
          example: "TO THE HONORABLE COURT: Defendant hereby moves this Court to dismiss the above-entitled action for lack of subject matter jurisdiction..."
        metadata:
          $ref: '#/components/schemas/DocumentMetadata'
        highlights:
          type: object
          description: Highlighted search terms
          additionalProperties:
            type: array
            items:
              type: string
        score:
          type: number
          format: float
          description: Relevance score
          example: 0.95
        storage_url:
          type: string
          format: uri
          description: URL to access the document
          example: "https://nyc3.digitaloceanspaces.com/motion-index-docs/documents/doc_123456.pdf"
      required:
        - document_id
        - filename
        - metadata
    DocumentMetadata:
      type: object
      properties:
        court:
          type: string
          example: "Superior Court of California, County of San Francisco"
        judge:
          type: string
          example: "Hon. Jane Smith"
        case_number:
          type: string
          example: "CV-2024-001234"
        case_name:
          type: string
          example: "People v. Defendant"
        filing_date:
          type: string
          format: date
          example: "2024-01-15"
        document_type:
          type: string
          example: "Motion"
        category:
          type: string
          example: "Pretrial Motion"
        parties:
          type: array
          items:
            type: string
          example: ["People of the State of California", "John Doe (Defendant)"]
        attorneys:
          type: array
          items:
            type: string
          example: ["Jane Attorney (Defense)", "Bob Prosecutor (DA)"]
        tags:
          type: array
          items:
            type: string
          example: ["jurisdiction", "motion to dismiss", "pretrial"]
        confidential:
          type: boolean
          default: false
        redacted:
          type: boolean
          default: false
...
    DocumentResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                document_id:
                  type: string
                  example: "doc_123456"
                filename:
                  type: string
                  example: "motion_to_dismiss.pdf"
                content_type:
                  type: string
                  example: "application/pdf"
                file_size:
                  type: integer
                  description: File size in bytes
                  example: 1048576
                metadata:
                  $ref: '#/components/schemas/DocumentMetadata'
                storage_url:
                  type: string
                  format: uri
                  example: "https://nyc3.digitaloceanspaces.com/motion-index-docs/documents/doc_123456.pdf"
                created_at:
                  type: string
                  format: date-time
                  example: "2024-01-15T10:30:00Z"
                updated_at:
                  type: string
                  format: date-time
                  example: "2024-01-15T10:30:00Z"
              required:
                - document_id
                - filename
                - content_type
                - metadata
...
    DocumentStatsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                total_documents:
                  type: integer
                  example: 15432
                total_size_bytes:
                  type: integer
                  example: 8934567234
                document_types:
                  type: object
                  additionalProperties:
                    type: integer
                  example:
                    Motion: 5678
                    Brief: 3456
                    Order: 2345
                courts:
                  type: object
                  additionalProperties:
                    type: integer
                  example:
                    "Superior Court of California": 12345
                    "Court of Appeal": 2345
                date_range:
                  type: object
                  properties:
                    earliest:
                      type: string
                      format: date
                      example: "2020-01-01"
                    latest:
                      type: string
                      format: date
                      example: "2024-01-15"
                last_updated:
                  type: string
                  format: date-time
                  example: "2024-01-15T10:30:00Z"
              required:
                - total_documents
                - total_size_bytes
                - document_types
                - courts
                - date_range
                - last_updated
    MetadataFieldsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                fields:
                  type: object
                  additionalProperties:
                    type: object
                    properties:
                      type:
                        type: string
                        enum: [string, integer, float, boolean, date, array]
                      searchable:
                        type: boolean
                      filterable:
                        type: boolean
                      values:
                        type: array
                        items:
                          type: string
                        description: Available values for enum fields
                      range:
                        type: object
                        properties:
                          min:
                            type: string
                          max:
                            type: string
                        description: Range for numeric/date fields
              required:
                - fields
...